Fixing compiler errors with GBFS and player + playing test song

+++ src/utils/gbfs/gbfs.h
@@ -51,26 +51,28 @@ extern "C" {
       "PinEightGBFSSpace-" #filename "-" #kbytes;
 
 typedef struct GBFS_FILE {
-  char magic[16];   /* "PinEightGBFS\r\n\032\n" */
-  u32 total_len;    /* total length of archive */
-  u16 dir_off;      /* offset in bytes to directory */
-  u16 dir_nmemb;    /* number of files */
-  char reserved[8]; /* for future use */
+  char magic[16];           /* "PinEightGBFS\r\n\032\n" */
+  unsigned long total_len;  /* total length of archive */
+  unsigned short dir_off;   /* offset in bytes to directory */
+  unsigned short dir_nmemb; /* number of files */
+  char reserved[8];         /* for future use */
 } GBFS_FILE;
 
 typedef struct GBFS_ENTRY {
-  char name[24];   /* filename, nul-padded */
-  u32 len;         /* length of object in bytes */
-  u32 data_offset; /* in bytes from beginning of file */
+  char name[24];             /* filename, nul-padded */
+  unsigned long len;         /* length of object in bytes */
+  unsigned long data_offset; /* in bytes from beginning of file */
 } GBFS_ENTRY;
 
 const GBFS_FILE* find_first_gbfs_file(const void* start);
 const void* skip_gbfs_file(const GBFS_FILE* file);
-const void* gbfs_get_obj(const GBFS_FILE* file, const char* name, u32* len);
+const void* gbfs_get_obj(const GBFS_FILE* file,
+                         const char* name,
+                         unsigned long* len);
 const void* gbfs_get_nth_obj(const GBFS_FILE* file,
                              size_t n,
                              char* name,
-                             u32* len);
+                             unsigned long* len);
 void* gbfs_copy_obj(void* dst, const GBFS_FILE* file, const char* name);
 size_t gbfs_count_objs(const GBFS_FILE* file);
